HBASE-25807 Move method reportProcedureDone from RegionServerStatus.proto to Master.proto
